Refactor Address, Email, Name, Phone from person to contact
