+++ src/test/java/seedu/address/model/person/EmailTest.java
@@ -5,6 +5,7 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
+import seedu.address.model.contact.Email;
 import seedu.address.testutil.Assert;
 
 public class EmailTest {
